ANDROID: tpm: Restore quality of tpm-rng to 1000
This change reapplies the applicable portion of
commit d0621b4cd498 ("CHROMIUM: tpm-rng: Make hwrng working")

The idea is to set a non-zero quality in the tpm-rng, so that
the hwrng_fill background thread gets started. This thread is able
to fill the randomness pool in the background, which helps speed up
boot time.  The original change contains a bit more context.

Bug: chromium:673183
Bug: 78841590
Bug: 356779145
Upstream-Task: 400625434
Test: Build kernel for Cheza, though no TPM yet.
Test: Deploy grunt kernel, UI comes up to login screen in < 15 seconds.

Change-Id: I63a4009593eb5cec655ccc9ccacefccf3b076da5
Signed-off-by: Evan Green <evgreen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/967183
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Signed-off-by: Mateusz Urbańczyk <urbanczyk@google.com>