ANDROID: pkvm: x86: Introduce vEPT to record guest EPT information
KVM-high manages a guest EPT for its nested VM, which shall be shadowed
to shadow EPT in pKVM hypervisor.

Introduce a per vcpu pgtable(vEPT) to save the guest EPT related
information from KVM-high when vmwrite(EPTP).

Each time when a vcpu vmwrites to EPTP, pKVM will setup the vEPT
according to the new value or just do nothing if it is the same
with the last one.

Bug: 395299836
Test: Boot, verify cpus are de-privileged and run a minimal protected vm.

Change-Id: I1d07c309c203ba4e5d9b23d71915a08f50c2d187
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>
Signed-off-by: Vineeth Pillai <vineethrp@google.com>