ANDROID: pkvm: x86: add shadow EPT invalidation & destroy support for protected VM
For protected VM, emulating the invept(shadow EPT invalidation) by setting
the leaf entry as invalid by clearing present bits while keeping the page
state and phys address still in the page entry. Doing so because pkvm cannot
undoate the memory back to host, and pkvm also cannot allow changing the
mapping of a GPA from one page to another. With setting this as non-presented
while still keeping the page state and physical address in, pkvm can re-create
this mapping with new property bits required and also can detect if the phys
address for this GPA is is changed or not.

And for shadow EPT destroy, pkvm will undonate all pages whose page state
showed owned this page in the page table entry, and before return the page
to host, the page shall be memset to 0 to avoid secret leakage.

Bug: 395299836
Test: Boot, verify cpus are de-privileged and run a minimal protected vm.

Change-Id: Icd42a44a3d53ec8b338f8b110d07b9d2a4b5ef02
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>
Signed-off-by: Vineeth Pillai <vineethrp@google.com>