ANDROID: pkvm: x86: Add support for clear "Suppress #VE" bit in pkvm
Since we initialized every pte with "Suppress #VE" bit set, no #VE will
be triggered. Protected VM only expects to receive #VE when it accesses
the MMIO, so the pkvm needs to set the mmio pte as invalid and also clear
the "Suppress #VE" bit.

In this patch the pkvm provides a hypercall which is SET_MMIO_VE to the
KVM. KVM can use this hypercall to set a mmio pte with "Suppress #VE"
cleared in the shadow ept in pkvm.

TODO: Consider the scenario of changing a MMIO entry in EPT to memory
entry. If a device's ROM first needs some emulation when the first time
of accessing, VMM can make the accessing of this ROM trapped. After the
required emulation is done and no need emulation later, then VMM can
map this ROM to memory.

Bug: 395299836
Test: Boot, verify cpus are de-privileged and run a minimal protected vm.

Change-Id: Iec3cff8ae9eeda8dc1e0276ff1b3f38e012540b5
Signed-off-by: Shaoqin Huang <shaoqin.huang@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>
Signed-off-by: Vineeth Pillai <vineethrp@google.com>