CHROMIUM: snd: virtio: Do not call virtsnd_remove on probe failed
When virtsnd_probe failed (e.g. timeout), it calls virtsnd_remove in its
clean up path. We suspect that virtsnd_remove may not be able to clean
up the device in a partially initialized state and cause a deadlock.

virtsnd_remove is also used as the .remove callback, and may lead to a
double free if the device is removed after the failure.

Remove the virtsnd_remove from the probe clean up path.

BUG=b:409866129
TEST=No virtsnd_remove log with forced timeout in host implementation

Change-Id: I95593d93665bc39e14efc892360458b819cce262
Signed-off-by: Pattara Teerapong <pteerapong@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6496108
Reviewed-by: Judy Hsiao <judyhsiao@google.com>